+++ packages/ui/src/Autosuggest/components/Fullscreen/contexts.ts
@@ -1,0 +1,18 @@
+import {createContext, useContext} from 'react';
+
+interface FullscreenContextType {
+  handleClose: () => void;
+  handleFocus: () => void;
+  setInputRef: (arg: HTMLInputElement | HTMLTextAreaElement) => void;
+  value: string;
+}
+
+export const FullscreenContext = createContext<FullscreenContextType | null>(
+  null,
+);
+
+export const useFullscreenContext = (): FullscreenContextType => {
+  const context = useContext(FullscreenContext);
+  if (!context) throw new Error('Context "FullscreenContext" not found');
+  return context;
+};