+++ packages/ui/src/Autosuggest/components/Fullscreen/types.ts
@@ -1,0 +1,4 @@
+export type TWithoutAddons<
+  T extends {leading?: React.ReactNode; trailing?: React.ReactNode}
+> = Omit<T, 'leading' | 'trailing'>;
+export type TIconPosition = 'leading' | 'trailing' | 'none';