+++ packages/ui/src/Autosuggest/index.ts
@@ -1,0 +1,9 @@
+import Autosuggest from './components/Default';
+import type {AutosuggestProps} from './components/Default';
+
+import FullscreenAutosuggest from './components/Fullscreen';
+import type {FullscreenAutosuggestProps} from './components/Fullscreen';
+
+export {Autosuggest, FullscreenAutosuggest};
+
+export type {AutosuggestProps, FullscreenAutosuggestProps};